+++ application/controllers/Player_Portfolio.php
@@ -1,23 +1,36 @@
 <?php
-
+/**
+ * The player portfolio page shows the recent trading activity and current
+ * holdings for a specific player, with a dropdown to select any other player.
+ * Trading activity includes purchases and sales that your app brokered.
+ * Current holdings shows the quantity held for each stock on the BSX. If no
+ * player is specified, then the page shows the portfolio of the currently
+ * logged in user.
+ */
 class Player_Portfolio extends MY_Controller {
-    function __construct()
-    {
+    function __construct() {
         parent::__construct();
     }
 
+    /**
+     * Index page for this controller.
+     */
     function index() {
-        $this->load->model('StocksModel');
-
-        $players = $this->StocksModel->getPlayers();
-        foreach ($players as $row) {
-            echo $row['Player'];
-            echo $row['Cash'];
-        }
-
         $this->data['pagetitle'] = 'Player Portfolio';
         $this->data['pagebody'] = 'player_portfolio';
-        $this->data['players'] = $players;
+        $this->loadPlayerNamesToOptions();
         $this->render();
+    }
+
+    /**
+     * Query the StocksModel Players table. Set each player row as a dropdown option.
+     */
+    function loadPlayerNamesToOptions() {
+        $options = "";
+        $players = $this->StocksModel->getPlayers();
+        foreach ($players as $row) {
+            $options .= "<option value=" . $row['Player'] . ">" . $row['Player'] . "</option>";
+        }
+        $this->data['options'] = $options;
     }
 }